create sharedpreferencesmanager, refactor packages, do some not null assertions, use mainapp to write to database
