+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             </intent-filter>
         </activity>
         <receiver
-            android:name="com.edu.uni.augsburg.uniatron.stepcounter.BroadcastReceiverOnStateChanged"
+            android:name="com.edu.uni.augsburg.uniatron.broadcastreceiver.BroadcastReceiverOnStateChanged"
             android:enabled="true"
             android:exported="false">
             <intent-filter>
@@ -26,7 +26,7 @@
             </intent-filter>
         </receiver>
         <service
-            android:name=".stepcounter.StepCountService"
+            android:name=".service.StepCountService"
             android:label="StepCountService" >
         </service>
     </application>
